move sorting functions to compatibility
In addition, the traversals are capable of a lazy sort
which should be faster when ties can be broken with a simple
count_ops comparison.

Also added permutations to iterables imports.
