+++ sympy/printing/ccode.py
@@ -12,7 +12,7 @@ source code files that are compilable without further modifications.
 from sympy.core import S, C
 from sympy.printing.codeprinter import CodePrinter
 from sympy.printing.precedence import precedence
-from sympy.utilities.misc import default_sort_key
+from sympy.core.compatibility import default_sort_key
 
 # dictionary mapping sympy function to (argument_conditions, C_function).
 # Used in CCodePrinter._print_Function(self)
