+++ sympy/printing/latex.py
@@ -12,7 +12,7 @@ from sympy.core.sympify import SympifyError
 import sympy.mpmath.libmp as mlib
 from sympy.mpmath.libmp import prec_to_dps
 
-from sympy.utilities.misc import default_sort_key
+from sympy.core.compatibility import default_sort_key
 from sympy.utilities.iterables import has_variety
 
 import re
