+++ sympy/solvers/recurr.py
@@ -59,7 +59,7 @@ from sympy.polys import Poly, quo, gcd, lcm, roots, resultant
 from sympy.functions import binomial, factorial, FallingFactorial, RisingFactorial
 from sympy.matrices import Matrix, casoratian
 from sympy.concrete import product
-from sympy.utilities.misc import default_sort_key
+from sympy.core.compatibility import default_sort_key
 from sympy.utilities.iterables import numbered_symbols
