+++ sympy/utilities/__init__.py
@@ -5,7 +5,7 @@ from iterables import (flatten, group, take, subsets,
     variations, numbered_symbols, cartes, capture, dict_merge,
     postorder_traversal, interactive_traversal,
     prefixes, postfixes, sift, topological_sort, unflatten,
-    quick_sort, has_dups, has_variety, reshape)
+    quick_sort, has_dups, has_variety, reshape, default_sort_key)
 
 from lambdify import lambdify
 from source import source
@@ -16,5 +16,3 @@ from runtests import test, doctest
 
 from cythonutils import cythonized
 from timeutils import timed
-
-from misc import default_sort_key
