+++ sympy/utilities/iterables.py
@@ -3,16 +3,14 @@ import random
 from operator import gt
 
 from sympy.core import Basic, C
-from sympy.core.compatibility import is_sequence, iterable  # logical location
-from sympy.core.compatibility import (as_int, quick_sort,
-         product as cartes, combinations, combinations_with_replacement)
-from sympy.utilities.misc import default_sort_key
 from sympy.utilities.exceptions import SymPyDeprecationWarning
 
 # this is the logical location of these functions
-from sympy.utilities.misc import default_sort_key
-from sympy.core.compatibility import (is_sequence, iterable,
-    product as cartes, combinations, combinations_with_replacement)
+from sympy.core.compatibility import (
+    as_int, big_first_keys, combinations, combinations_with_replacement,
+    default_sort_key, is_sequence, iterable, lazyDSU_sort, permutations,
+    product as cartes, quick_sort, small_first_keys
+)
 
 
 def flatten(iterable, levels=None, cls=None):
@@ -202,8 +200,9 @@ def postorder_traversal(node, key=None):
 
     Examples
     ========
-    >>> from sympy import symbols, default_sort_key
-    >>> from sympy.utilities.iterables import postorder_traversal
+
+    >>> from sympy.utilities.iterables import (
+    ...    postorder_traversal, big_first_keys, default_sort_key)
     >>> from sympy.abc import w, x, y, z
 
     The nodes are returned in the order that they are encountered unless key
@@ -213,6 +212,8 @@ def postorder_traversal(node, key=None):
     [z, y, x, x + y, z*(x + y), w, w + z*(x + y)]
     >>> list(postorder_traversal(w + (x + y)*z, key=default_sort_key))
     [w, z, x, y, x + y, z*(x + y), w + z*(x + y)]
+    >>> list(postorder_traversal(w + (x + y)*z, key=big_first_keys))
+    [x, y, x + y, z, z*(x + y), w, w + z*(x + y)]
 
 
     """
@@ -220,7 +221,10 @@ def postorder_traversal(node, key=None):
         args = node.args
         if key:
             args = list(args)
-            args.sort(key=key)
+            if isinstance(key, (list, tuple)):
+                args = lazyDSU_sort(args, key)
+            else:
+                args.sort(key=key)
         for arg in args:
             for subtree in postorder_traversal(arg, key):
                 yield subtree
@@ -384,6 +388,7 @@ def subsets(seq, k=None, repetition=False):
 
        Examples
        ========
+
        >>> from sympy.utilities.iterables import subsets
 
        subsets(seq, k) will return the n!/k!/(n - k)! k-subsets (combinations)
@@ -1091,6 +1096,7 @@ def has_dups(seq):
 
     Examples
     ========
+
     >>> from sympy.utilities.iterables import has_dups
     >>> from sympy import Dict, Set
 
@@ -1112,6 +1118,7 @@ def has_variety(seq):
 
     Examples
     ========
+
     >>> from sympy.utilities.iterables import has_variety
     >>> from sympy import Dict, Set
 
@@ -1358,90 +1365,6 @@ def generate_oriented_forest(n):
                     break
             else:
                 break
-
-# define keys for use with lazyDSU_sort; define as tuples
-# so they can't be inadvertently changed
-
-small_first_keys = (lambda _: _.count_ops(), default_sort_key)
-big_first_keys = (lambda _: -(_.count_ops()), default_sort_key)
-
-def lazyDSU_sort(seq, keys, warn=False):
-    """Return sorted seq, breaking ties by applying keys only when needed.
-
-    If ``warn`` is True then an error will be raised if there were no
-    keys remaining to break ties.
-
-    Examples
-    ========
-
-    >>> from sympy.utilities.iterables import lazyDSU_sort, default_sort_key
-    >>> from sympy.abc import x, y
-
-    The count_ops is not sufficient to break ties in this list and the first
-    two items appear in their original order: Python sorting is stable.
-
-    >>> lazyDSU_sort([y + 2, x + 2, x**2 + y + 3],
-    ...    [lambda x: x.count_ops()], warn=False)
-    ...
-    [y + 2, x + 2, x**2 + y + 3]
-
-    The use of default_sort_key allows the tie to be broken (and warn can
-    be safely left False).
-
-    >>> lazyDSU_sort([y + 2, x + 2, x**2 + y + 3],
-    ...    [lambda x: x.count_ops(),
-    ...     default_sort_key])
-    ...
-    [x + 2, y + 2, x**2 + y + 3]
-
-    Here, sequences are sorted by length, then sum:
-
-    >>> seq, keys = [[[1, 2, 1], [0, 3, 1], [1, 1, 3], [2], [1]], [
-    ...    lambda x: len(x),
-    ...    lambda x: sum(x)]]
-    ...
-    >>> lazyDSU_sort(seq, keys, warn=False)
-    [[1], [2], [1, 2, 1], [0, 3, 1], [1, 1, 3]]
-
-    If ``warn`` is True, an error will be raised if there were not
-    enough keys to break ties:
-
-    >>> lazyDSU_sort(seq, keys, warn=True)
-    Traceback (most recent call last):
-    ...
-    ValueError: not enough keys to break ties
-
-
-    Notes
-    =====
-
-    The decorated sort is one of the fastest ways to sort a sequence for
-    which special item comparison is desired: the sequence is decorated,
-    sorted on the basis of the decoration (e.g. making all letters lower
-    case) and then undecorated. If one wants to break ties for items that
-    have the same decorated value, a second key can be used. But if the
-    second key is expensive to compute then it is inefficient to decorate
-    all items with both keys: only those items having identical first key
-    values need to be decorated. This function applies keys successively
-    only when needed to break ties.
-    """
-    d = defaultdict(list)
-    keys = list(keys)
-    f = keys.pop(0)
-    for a in seq:
-        d[f(a)].append(a)
-    seq = []
-    for k in sorted(d.keys()):
-        if len(d[k]) > 1:
-            if keys:
-                d[k] = lazyDSU_sort(d[k], keys, warn)
-            elif warn:
-                raise ValueError('not enough keys to break ties')
-            seq.extend(d[k])
-        else:
-            seq.append(d[k][0])
-    return seq
-
 
 def minlex(seq, directed=True):
     """Return a tuple where the smallest element appears first; if
